parking-lots-architecture/requirements.txt
Flask==1.1.1
h5py==2.10.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
matplotlib==3.0.3
mysql-connector-python==8.0.19
numpy==1.18.2
paho-mqtt==1.4.0
pandas==0.25.3
plotly==4.4.1
PyYAML==3.13
requests==2.21.0
SQLAlchemy==1.3.13
sshtunnel==0.1.5
tensorboard==2.2.1
tensorflow==2.2.1
tensorflow-estimator==2.2.0
